requirements.txt
python-telegram-bot==20.6
pybit==5.10.1
python-dotenv